fix: optimizeCues O(n^2)
