+++ src/web/cues.ts
@@ -1,7 +1,6 @@
 import { getAttribute } from './attribute';
 import { getElementText } from './element';
 import { isDynamic } from './isDynamic';
-import { SelectorPart } from './types';
 
 const DEFAULT_ATTRIBUTE_LIST =
   'data-cy,data-e2e,data-qa,/^data-test.*/,/^qa-.*/';
@@ -274,33 +273,12 @@ export const buildCues = ({ attribute, isClick, target }: BuildCues): Cue[] => {
   return cues;
 };
 
-export const buildSelectorParts = (cues: Cue[]): SelectorPart[] => {
-  const levels = [...new Set(cues.map((cue) => cue.level))];
-
-  // sort descending
-  levels.sort((a, b) => b - a);
-
-  const parts: SelectorPart[] = [];
-
-  levels.forEach((level) => {
-    const cuesForLevel = cues.filter((cue) => cue.level === level);
-
-    const textCues = cuesForLevel.filter((cue) => cue.type === 'text');
-    if (textCues.length) {
-      parts.push({ name: 'text', body: textCues[0].value });
-      return;
+export const findNearestPreferredAttributeCue = (cues: Cue[]): Cue | null => {
+  return cues.reduce((foundCue, cue) => {
+    if (cue.penalty === 0 && (!foundCue || foundCue.level > cue.level)) {
+      return cue;
     }
 
-    cuesForLevel.sort((a, b) => {
-      if (a.type === 'tag') return -1;
-      if (b.type === 'tag') return 1;
-      return 0;
-    });
-
-    const bodyValues = cuesForLevel.map((cue) => cue.value);
-
-    parts.push({ name: 'css', body: bodyValues.join('') });
-  });
-
-  return parts;
+    return foundCue;
+  }, null);
 };
